ISDB-320: Refactoring
